Migrate api-key OpenAI provider to the Responses API
gpt-5.x reject function tools + reasoning_effort on Chat Completions, so route
the api-key `openai` provider through the Responses API instead. The
OpenAI-compatible providers (together, groq, fireworks, xai, llama, dashscope)
keep the Chat Completions OpenAIProvider.

- Add OpenAIResponsesProvider (api key against api.openai.com/v1/responses) and
  extract shared Responses machinery (request building, streaming, reasoning
  continuity, token counting) into ResponsesProviderBase in responses_common.
- Refactor ChatGPTOAuthProvider to build on ResponsesProviderBase; only the
  transport (OAuth, backend URL, default model) differs.
- Update openai model specs to openai_responses_reasoning mode, drop temperature,
  and use "minimal" effort instead of "none".
- Stamp messages with the real provider (openai vs openai_chatgpt) and allow
  reasoning to replay across the two Responses backends via compatibility groups.
- Add openai_chatgpt to the instrumented client usage-tracking provider lists.
